% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootCor}
\alias{bootCor}
\title{Bootstrap correlation coefficients from two vectors of values.}
\usage{
bootCor(X, Y, size = 10000)
}
\arguments{
\item{X}{vector of values}

\item{Y}{vector of values (same length as X)}

\item{size}{number of resampling iterations}
}
\value{
dataframe of estimated correlations, estimated p.values, and the
observed correlation in the data.
}
\description{
Bootstrap correlation coefficients from two vectors of values.
}
