% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootSummary}
\alias{bootSummary}
\title{Bootstrap statistical summaries}
\usage{
bootSummary(x, size = 10000)
}
\arguments{
\item{x}{vector of observations}

\item{size}{Number of bootstrapping iterations}
}
\value{
Dataframe with summary statistics for bootstrapped median and .05/.95 quantiles
of the input observations
}
\description{
Bootstrap statistical summaries
}
