% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_helpers.R
\name{describe}
\alias{describe}
\title{Simple descriptor function}
\usage{
describe(df, method = "classical", ...)
}
\arguments{
\item{df}{data frame object}

\item{method}{"classical" or "bootstrap"}

\item{...}{...}
}
\value{
ans a named vector of statistical values
}
\description{
Simple descriptor function
}
