% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAT.R
\name{dScore}
\alias{dScore}
\title{Calculate a D score based on vectors of means, SDs, and Ns}
\usage{
dScore(m1, m2, sd1, sd2, n1, n2)
}
\arguments{
\item{m1}{mean of block 1}

\item{m2}{mean of block 2}

\item{sd1}{standard deviation of block 1}

\item{sd2}{standard deviation of block 2}

\item{n1}{n (count) of block 1}

\item{n2}{n (count) of block 2}
}
\value{
vector of D scores
}
\description{
Calculate a D score based on vectors of means, SDs, and Ns
}
