% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAT.R
\name{expt_score.iat}
\alias{expt_score.iat}
\title{Score IAT}
\usage{
\method{expt_score}{iat}(df, excludedSubjects = NA, platform = "inquisit",
  ...)
}
\arguments{
\item{df}{Experimental dataframe.  Must contain columns subject,}

\item{platform}{experimental platform (default = "inquisit")}
}
\value{
a list of data.frames
}
\description{
Processes IAT experimental data and returns D score and
appropriate statistics
}
