% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSS.R
\name{expt_score.pss}
\alias{expt_score.pss}
\title{Score a Probabalistic Stimulus Selection task}
\usage{
\method{expt_score}{pss}(df, platform = "eprime")
}
\arguments{
\item{df}{a dataframe}

\item{platform}{experimental platform}
}
\value{
results
}
\description{
Score a Probabalistic Stimulus Selection task
}
