% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{expt_score}
\alias{expt_score}
\title{Generic scoring function helper}
\usage{
expt_score(df, excludedSubjects, platform, ...)
}
\arguments{
\item{df}{}
}
\value{
NA = referrer function
}
\description{
Generic scoring function helper
}
