% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simon.R
\name{expt_score.simon}
\alias{expt_score.simon}
\title{Score Simon Task}
\usage{
\method{expt_score}{simon}(df, excludedSubjects, platform, ...)
}
\arguments{
\item{df}{raw dataframe from experimental platform}

\item{platform}{specify experimental platform (default = "inquisit")}
}
\value{
scored dataframe
}
\description{
Score Simon Task
}
