% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{expt_setup}
\alias{expt_setup}
\title{Setup the experiment object returned by scoring routines}
\usage{
expt_setup(df, excludedSubjects, platform)
}
\arguments{
\item{df}{raw experimental data frame}

\item{excludedSubjects}{vector of excluded subjects}

\item{platform}{experimental platform}
}
\value{
expt experimental object for further processing
}
\description{
Setup the experiment object returned by scoring routines
}
