% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAT.R, R/iat.score.R
\name{iat_findOrder}
\alias{iat_findOrder}
\alias{iat_findOrder}
\title{Find compatibility order of IAT}
\usage{
iat_findOrder(df)

iat_findOrder(df)
}
\arguments{
\item{df}{same as scoring frame}

\item{df}{same as scoring frame}
}
\value{
data frame of subjects with order

data frame of subjects with order
}
\description{
Find compatibility order of IAT

Find compatibility order of IAT
}
