% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAT.R
\name{iat_flagError}
\alias{iat_flagError}
\title{Flag Error Rate (IAT)}
\usage{
iat_flagError(df, criterion)
}
\arguments{
\item{df}{dataframe containing subject number, pairing, and correct (1/0 or T/F)}

\item{criterion}{maximum acceptable error rate}
}
\value{
dataframe of subject numbers and flag values (true/false)
}
\description{
Flag Error Rate (IAT)
}
