% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAT.R, R/iat.flag.R
\name{iat_flagSpeed}
\alias{iat_flagSpeed}
\alias{iat_flagSpeed}
\title{Flag subjects for excessive speed}
\usage{
iat_flagSpeed(df, threshold = 0.3, criterion = 0.1)

iat_flagSpeed(df, threshold = 0.3, criterion = 0.1)
}
\arguments{
\item{df}{dataframe containing columns subject, pairing, and latency}

\item{threshold}{value (in seconds) to assess as a speed threshold (i.e. trials under 300ms)}

\item{criterion}{value (as proportion) above which to flag a subject for discard, i.e. .10 of trials under 300ms}

\item{df}{dataframe containing columns subject, pairing, and latency}

\item{threshold}{value (in seconds) to assess as a speed threshold (i.e. trials under 300ms)}

\item{criterion}{value (as proportion) above which to flag a subject for discard, i.e. .10 of trials under 300ms}
}
\value{
a vector of subjects, proportion under threshold, and flag (logical)

a vector of subjects, proportion under threshold, and flag (logical)
}
\description{
Flag subjects for excessive speed

Flag subjects for excessive speed
}
