% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{metaBootCor}
\alias{metaBootCor}
\title{Bootstrap correlation coefficients from two vectors of values.}
\usage{
metaBootCor(X, Y, size = 10000)
}
\arguments{
\item{X}{vector of values}

\item{Y}{vector of values (same length as X)}

\item{size}{number of resampling iterations}
}
\value{
Observed correlation (from sample) and 5% and 95% quantiles
}
\description{
Bootstrap correlation coefficients from two vectors of values.
}
