% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{parseEdat}
\alias{parseEdat}
\title{Parse ePrime .txt logfile}
\usage{
parseEdat(filename)
}
\arguments{
\item{filename}{of ePrime logfile.txt}
}
\value{
a dataframe from the experiment, with subject number, session number, and date
}
\description{
Parse ePrime .txt logfile
}
