% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simon.R
\name{print.summary.simon}
\alias{print.summary.simon}
\title{Print method for Simon experiments}
\usage{
\method{print}{summary.simon}(x, ...)
}
\arguments{
\item{x}{input simon object}

\item{...}{NA}
}
\value{
NA
}
\description{
Print method for Simon experiments
}
