% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{read_merge}
\alias{read_merge}
\title{Read and merge files of a particular type, useful when individual subject
files are saved separately by the experimental software}
\usage{
read_merge(pattern, ...)
}
\arguments{
\item{pattern}{a regular expression pattern to describe files to import.  Should include
information about the filename as well as its extension}

\item{...}{additional parameters such as column specifications for read_tsv}
}
\value{
a single merged files
}
\description{
Read and merge files of a particular type, useful when individual subject
files are saved separately by the experimental software
}
\examples{
pattern="ageiat_raw+.*iqdat"
}
