% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{scoreExpt}
\alias{scoreExpt}
\title{Score Experiment}
\usage{
scoreExpt(df, type, excludedSubjects = NA, platform = "default", ...)
}
\arguments{
\item{df}{Dataframe containing experimental data produced by an experiment presentation software}

\item{type}{Type of experiment, e.g. "simon" "IAT"}

\item{...}{}
}
\value{
NA - referrer function
}
\description{
Score Experiment
}
