% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAT.R
\name{scoreIAT}
\alias{scoreIAT}
\title{Score IAT}
\usage{
scoreIAT(df, platform = "inquisit")
}
\arguments{
\item{df}{Experimental dataframe.  Must contain columns subject,}

\item{platform}{experimental platform (default = "inquisit")}

\item{excludedsubjects}{a vector of excluded subject numbers}
}
\value{
a list of data.frames
}
\description{
Processes IAT experimental data and returns D score and
appropriate statistics
}
