% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simon.R
\name{summary.simon}
\alias{summary.simon}
\title{Summary method for Simon task}
\usage{
\method{summary}{simon}(x, statmethod = "classical", ...)
}
\arguments{
\item{x}{experiment object to be summarized}

\item{method}{method to use for confidence intervals - "classical" or "bootstrap"}
}
\value{
NA
}
\description{
Summary method for Simon task
}
